Update the type from int to uint64

+++ internal/sysinfo/memtotal_linux.go
@@ -2,7 +2,7 @@ package sysinfo
 
 import "os"
 
-func PhysicalMemoryBytes() (int, error) {
+func PhysicalMemoryBytes() (uint64, error) {
 	f, err := os.Open("/proc/meminfo")
 	if err != nil {
 		return 0, err
